+++ run_dense_hyper.py
@@ -219,8 +219,8 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="Convolutional Bayesian Neural Networks for CIFAR-10")
     parser = argparse.ArgumentParser(description="Convolutional Bayesian Neural Networks for CIFAR-10")
     parser.add_argument("--train_index", type=float, default=50000)
-    parser.add_argument("--num_warmup", type=float, default=100)
-    parser.add_argument("--num_samples", type=float, default=100)
+    parser.add_argument("--num_warmup", type=int, default=100)
+    parser.add_argument("--num_samples", type=int, default=100)
     parser.add_argument("--gpu", type=bool, default=False)
     args = parser.parse_args()
